Rename training to trainingdata
